refactor: Stylistic changes to event summary plot
Make some stylistic changes the the event summary plot. In this process,
the `event_summary` function was split into logical sections, which were
assigned to subroutines.

Changes include:

- Fix legend position
- Add magnitude information in text summary
- Triangles -> bars for predicted traveltimes. Colour scheme consistent with
the phase pick plots
- Changed the colours used for the waveforms from the default set to a less
'in-your-face' trio. Orange for Z maps to the red P phase pick. Blue/green
for N/E maps to the blue S phase pick.
- Thinned out lines
- Added some alpha
- Spaced the waveforms so as not to overlap (previously they were spaced every
1 unit in y - now spaced every 2 units in y)
- Set the x limit to be 0.1 s before the origin time to 1.5 x the maximum
predicted traveltime
- The station traces are normalised for a window from the event origin time
to 1.5 * the maximum traveltime -for that station-. Should help when there are
several events closely spaced in time.
- Add bounding-boxes to data trace tags.
- Make use of the custom DateFormatter introduced in 29012bf

In general, I have also tried to minimise the number of places one needs
to change plot variables (fontsize, etc) using rcParam as a context
manager. This is a precursor for custom stylesheets.