refactor: Finalise divorce between STALTAOnset and GaussianPicker
Cut the final explicit tie between STALTAOnset and GaussianPicker - the
interface is now fully abstracted such that users can use the GaussianPicker
module with custom Onsets.

Added a gaussian_halfwidth method to the Onset base class that will raise an
error if the user tries to use the GaussianPicker without implementing that
part of the interface.
Added a gaussian_halfwidth method to the STALTAOnset class that takes a
phase argument and returns the phase-relevant short-term average window
length (in terms of samples).