refactor: Stylistic changes to trigger summary plot
Make some stylistic changes the the trigger summary plot. In this process,
the `trigger_summary` function was split into logical sections, which were
assigned to subroutines.

Changes include:

- Add time span information to summary text plot
- Make use of the custom DateFormatter introduced in 29012bf
- Move the colorbar to a less prominent position in the plot
- Zoomed in on the y axis of the station availability plot to only show
a window +10% of the max / -20% of the min station availability.
- Added bounding-boxes to the data trace plots.

NOTE: Will need to manually merge the changes introduced in #80 to the naming
of `minimum_repeat`.