fix: Resolve issue with tight layout
There was an issue with using tight layout to tidy up and automatically
ascertain how best to group the plots and minimise the horizontal/vertical
spacing between them. The error message was:

UserWarning: Tight layout not applied. tight_layout cannot make axes height
small enough to accommodate all axes decorations
    fig.tight_layout(pad=1, h_pad=0)

This was being caused by station annotations not being clipped to the plot
bounds (which may or may not be a bug in matplotlib itself). This was
resolved by explicitly setting the keyword 'clip_on=True' in the
call to 'annotation'.

Fingers crossed we never see this again.