feat(lut.plot): Make x and y scale equal on LUT plots
Previously the LUT map plot aspect ratio was determined by the bounding box. Now it is
scaled by the grid size such that the x and y scale are the same.

This required a bit of faff, because matplotlib doesn't have a sensible way (as far
as I can find) to share changes in subplot size between subplots with shared axes. So
I did it very manually.

This has the small additional benefit that the two cross sections now have exactly equal
heights (in the depth direction) - previously the accuracy was constrained by gridspec
maths.