feat(LUT.plot): Add scalebar to map plots
Following on from making the x and y scales consistent, a scalebar has now been
added. Its length is chosen to be approximately 10% of the grid x dimension, but
always a whole number of cells wide. This should ensure sensible distances.

Unforunately the scalebar has to be fixed to a specific location. It has
been placed in the bottom right corner in the hope that this will be out of the
way most of the time.