fix: Minor LUT plotting bug
Fix a minor bug with plotting cross-sections of the LUT. Previously, the
grid was defined using the extent of the grid 'nodes'. This had a two-fold
effect on the plotted cross-sections of the coalescence volume.

1. The coalescence values were offset from the grid nodes by ~half a cell size
2. There were n grid cells being plotted in the space of n-1. This had a very
small, but noticeable effect on the size of each cell of the gridmesh used
in pcolormesh.

This was addressed by slightly altering the definition of the 'grid_extent'
property of the Grid3D class. Now it can return either the convex hull of
the grid nodes (default) or the convex hull of the grid of cells centred on
the grid nodes.