refactor: Semantic change to LUT attributes
Make a semantic change to the LUT attributes 'cell_count' and 'cell_size'.
These attributes have been mapped to 'node_count' and 'node_spacing', which
we feel better reflects what they actually mean. The grid is defined as a
series of nodes spaced by [x, y, z]. It is only when visualising the grid
that the notion of cells is helpful. This change additionally brings us in
line with the grid definition used by other softwares, e.g. NonLinLoc.

Added deprecation handlers and expanded the `update_lut` function to also
handle this change.

All tests pass, both using the old-style LUT (with deprecation handlers)
and creating a new LUT under the new definition.