chore(package-wide): Rename the package to PEP8 guidelines
Rename the package from 'QMigrate' -> 'quakemigrate' to conform to
the PEP8 style-guide on package names.

"Python packages should have short, all-lowercase names."

This commit will be followed by a series of commits bringing the package
up to date. These changes will include:

- imports
- docstrings
- Test scripts
- Update setup.py
- Update documentation
- Update examples

This will (obviously) cause a Breaking Change for users. This is unfortunately
unavoidable, but better to do this now.