feat: Windows compatibility and depency control
Add some minor tweaks to allow for Windows compatibility, namely:

- Custom build extension class to handle export symbols
- Pass a string object (rather than Path object) to CDLL
- Split the .c file into a .h header and a .c source code file and
dropped the explicit setting of export symbols.

In addition to these changes, handle various minor dependency changes:

- Recent matplotlib release (v3.3.0) causing some issues with plotting
-> solution has been to pin the version for now.
- Make scikit-fmm an optional install. This was causing some issues, and
is not always needed. This can be installed separately to the main
environment if required.
- Added a catch for the version of PROJ used as the backend for pyproj
to handle a known issue with v6.2.0 (see installation documentation for
more details)