chore: Update README, logo, docs
Update the README to give a clean, simple overview of the project. This
will be expanded slightly following a review of the language used to
describe QuakeMigrate. There will also be some badges added related to
Continuous Integration bots in the future.

Updated to use the new logo.

Add a CONTRIBUTING.md file that gives some guidelines for contributing
to QuakeMigrate, including a notice on the license and a Code of
Conduct.