fix(scan.py): Fix catch for case mags pre_pad < migrate pre_pad
Due to a poorly written if statement, the case that the pre_pad set for magnitudes is
smaller than that needed for migration in locate was not caught unless a pre_cut was
also set, resulting in a util.DataGapException for every event run in locate, and
no results.

This has been resolved.