fix(scan/scanmseed): Fix bug in `ScanmSEED.empty()`
`ScanmSEED.empty()` did not take the unit conversion factor as an argument, and did not
pass it on to `ScanmSEED.append()`, resulting in a fatal TypeError when handling a
util.DataGapException or util.ArchiveEmptyException.

This has been resolved in scan.py (where the function is called) and scanmseed.py
(where the function is defined).

A minor linting issue related to the previous commit has also been tidied.