feat(onset/lib): Write onsets into C
Write the onset functions into C. This also includes a correction to the
implicit window positioning for the centred onset function. Previously,
this was not behaving as expected.

Timing the functions confirmed that using our own C extensions for the
STA/LTA onset functions is faster than the old functions (though, in
fairness to NumPy, it was pretty quick) by a factor of ~3.

Some future work could be done to multiprocess the onset function
calculation.

In addition, the libraries were slightly renamed for clarity
('quakemigrate.c' -> 'migratelib.c'). Documentation was updated and the
names of variables used in migratelib.c for clarity.