fix(install): Address dependency issues
Add a pin for the pandas version (< 1.1) in order to avoid conflicts
with the recent changes made to matplotlib's datetime epoch.

Also resolve to use setuptools - this is a key package, distributed with
miniconda and other Python installations to handle package distribution
and installation.

Added some additional information in the installation guide per feedback
from users.