refactor: Use traces over arrays
Switch to use streams/traces (and selects) over arrays (and indexing) -
this has several knock-on benefits:

1. Easy access to stations/component specific data when needed, without
having to assume all ordering is retained. This will be useful in the
future if we wish to use sub-sets of the available data for a scan run.
2. Can now access the ObsPy machinery for component selection! This will
be introduced properly in the next commit, but this will allow for
targeted, but generalised, component selection e.g. *[N,E] will select
all traces with channel names ending in N OR E.

As a result of this change, some computations were moved, such as when
the data availability is determined and the resampling of the data
before calculating onset functions.

Switched to using the batching style used in trigger for reading in
station data from the archive for consistency. I think it might also be
possible to specify the subdivision of time used (currently hardcoded at
one day = 86400 seconds) if data is stored in more heavily subdivided
files (e.g. hourly files, if using high sampling rate data such as that
used in cryoseismological applications).

Resample is now in the util folder, and operates on streams/traces.

Update plots to use traces over arrays for signal plotting.

This commit will be supplemented by the following commit which will
propagate these changes into the onset function calculation, as well as
reworking how onsets themselves are calculated.