refactor(onset): Make onsets phase agnostic
Make the STALTAOnset class phase agnostic. To do this, the filters and
windows lengths are now specified using dictionaries, with the
corresponding phase being used the key. A fairly large change as part of
this was to where the combination of onsets on the horizontal was
performed. It is now moved before the log is taken. I tested it with the
old ordering to ensure the other changes had had no effect - can confirm
everything passed.

Upsample and refactor have been moved from the Archive class to the
Onset class.

Rename the sampling_rate variable in QuakeScan to scan_rate. This is to
distinguish it from the sampling rate used for the onset data. In
future, we would like to be able to specify different values to these to
'rates'.

Onset data is now stored using the same structure as used in the lookup
table class - a nested dictionary with keys of stations, with subkeys of
phases.

Traveltimes are now served based on the specific stations requested -
this should both ensure the ordering is correct when served up into the
core C routines and reduce any redundant data stacking!