fix(gaussianpicker): Fix check that picktime is inside picking window
A check had already been added to ensure the time of the pick was within the picking
window, but it used the incorrect index to make this check. It was based on the
duration of the onset function used for the gaussian fit, so in cases where the onset
function was above the exceedence value at the edge of the pick window, the 1 sample
padding used would give 1 sample of leeway to this check.

This has been corrected so the indices of the pick window itself are used in the test.

A couple of relevant debug messages have also been added.