fix(onsetlib.c): Reproduce previous behaviour
Slightly adjust the new centred STA/LTA C implementation to reproduce
the previous behaviour. This entails positioning the STA and LTA windows
consecutively (e.g. the first sample of the STW is one sample AFTER the
last sample of the LTW), and assigning the STA/LTA to the last sample of
the LTW.

In the future, a gap may be introduced which would allow the user to
further separate the windows in time. This would in theory make them
more statistically independent, giving a more robust PDF.

This change has been theoretically validated on a simple dataset and
passes all benchmark tests.

Note:
LTW/A = Long Term Window/Average
STW/A = Short Term Window/Average