feat: Make availability/picker phase agnostic
Make the availability phase agnostic - everything is now ready to go. QM
can be run on single phase (by simply specifying only a single column
when building the LUT). Minor changes were made to how the availability
is handled. A dictionary is now constructed that contains the list of
stations, then for each station whether each phase requested is
available. This is output using column header "{station}.{phase}" when
writing to file. This is extensible to any number of phases and avoids
as much hardcoding as possible.

Another major change was made to break the Gaussian picking down into
stages. This was necessary to make it easier to make phase agnostic, but
should also be more clear what exactly is done during picking.

Summary of changes
- Traveltime grids are now served based on the availability information
provided.
- Onsets are also now served based on the availability information
provided. This guarantees that they are served up in exactly the right
order when passed through the core stacking algorithm.
- Restructure the GaussianPicker to be more streamlined/clearer and
phase agnostic.

Note: The pick plotting works, but it does not always behave well when
onsets or data is missing. This will be ironed out during the review
stage by me.

Note: There are likely numerous instances of incomplete documentation
for functions. Hopefully these will all be ironed out in review!