feat: Rename onset/picker
Rename the onset and picker base class modules and modules that contain
any children of these. It's already abundantly clear that the modules in
signal/onsets are indeed onset related, and 'base' is a clearer
indicator of being an abstract base class module than the previous
names.

Will also help when merging in changes from other branches with changes
to onset.