fix(gaussianpicker): Fix dict overwrite issue
Python dicts are mutable. Python does not implicitly copy anything.

When the Gaussian fitting failed (for any reason - no values of onset
exceeding the threshold, curve_fit breaking, peak of Gaussian being
outside window), a really useful default dictionary, stored as a class
variable (DEFAULT_GAUSSIAN_FIT), was used to provide some suitably
'wrong' defaults. However, this was being assigned in the _fit_gaussian
function to some variable - a reference to this variable (in itself a
reference back to the original class variable) was then stored in the
big gaussianfits dictionary. However, each subsequent failure  would
update the values in the class variable, which would then pop up
everywhere else a reference to that specific instance was found.

Solved by adding a .copy() to all instances of self.DEFAULT_GAUSSIAN_FIT

While annoying, this only affected the plots. It did, in fact, prove to
be useful by making it look like obvious picks were being missed, when
in fact this was a separate issue that will be addressed in a future
update.