fix(gaussianpicker): use event marginal window
Revert removal of marginal window from the calculation of the pick
window. This was mistakenly removed. The marginal window is now added to
the pick window, but it is sourced from the Event object. This removes
the redundant ability to set the marginal_window on the GaussianPicker
object, which could have lead to inconsistent usage between phase
picking and event location.

Additionally, remove some changes left in regarding the setting of the
noise threshold.