feat(trigger summary): Make plotting optional
Make the plotting of trigger summary figures optional. This can be
toggled in the same way as the summary figures in QuakeScan - e.g.
trigger.plot_trigger_summary = True.

When this option is turned off, no plotting is performed.

When this option is turned on, all plotting is performed and the
resultant figure is *always* saved. Previously, it was possible to
toggle whether it was saved or not. This would mean running the code
twice, even if satisfied after exploring the data in the interactive
plot. The keyword argument (provided when running the trig.trigger
function) 'savefig' has been changed to 'interactive_plot', as it is far
more explicit.

I don't feel it is necessary to deprecate anything. The default
behaviour remains the same.

Final change was to add a call to plt.close(fig) at the end of the
plotting script, which ensures everything is closed down and there
aren't anymore memory leaks (which can become an issue if running
trigger, with plotting on, for an extended period of time).

All example scripts have been updated to reflect the changes.