feat(dike example): Added a new example - the 2014 Holuhraun dike intrusion
This adds a new example use-case for QM: detecting volcano-tectonic earthquakes
induced during the 2014 Bardarbunga-Holuhraun dike intrusion. This current commit
includes a fairly bare-bones implementation, to ensure the example works as expected
across multiple systems. It will be expanded in future and a Jupyter Notebook
will be added to provide further description and reasoning behind the choice of
parameters etc.

Here the waveform data is not stored on GitHub, but there is an additional script to
download the waveform data and instrument response information from IRIS. This will
allow us to create a new benchmark covering magnitude calculation as well as the core
stages of QM. A slimmed-down version will be prepared for use with a CI service to
provide more timely warnings of compatibility issues.