fix(locate): catch reading triggered events issue
Catch and handle (by means of an explicit error message) an issue
whereby the user might set an invalid run name or time period for
locate, but the script would crash with a fairly impenetrable error.

All previous behaviour identical, now simply checks if any suitable
triggered events files were found during the while loop. If not, raise
the error. Perhaps it would be worth catching this and printing
something neater before exiting the script, but the net effect is the
same.