feat(dike example benchmark): Add dike example benchmark
The dike example has been run as a benchmark. Currently this uses the entire
10 minute run window for the benchmark - it may be more efficient to just use a single
event. It depends what we want this test to cover.

The test_benchmarks script has been updated; first to add this second benchmark, but
also to fix a couple of minor issues. It now uses the __file__ magic to allow the
script to be run from any directory, and an additonal merge is added to allow for
cut waveforms from many events to be included in the locate tests.