fix(stalta.py): Add check that the waveform data covers the timespan
In determining whether data for a particular station/phase is available or not,
the check that the data covers the entire requested timespan had been lost. This
commit restores that check. However, the rigid calculation of the taper to be
applied to the data means the code as curretly written will still break. The
following commit will address that by switching to the obspy functions for
filtering, tapering etc.