feat: added more advanced station naming 	Added a feature such that stations can be defined in the stations input file using their full ids (i.e. network, station, location, channel). This improves the ability for QM users to select station and channel combinations from larger data archives.
The only significant change to the input file is that users should define
"Station" at minimum in their station input files. QM will then assemble a
naming string using wildards. Users can also add addition columns ("Network",
"Location" and "Channel") which can include regex expressions.

As part of this feature I have removed most of the default naming schemes
(leaving only SeisComp3). Novice users will struggle to find find where
these are defined and I find it difficult to remember the string syntax. I
suggest users instead use the format strings functionality which is more
flexible and easier to understand. But as always, this can be put back in if
necessary

I have tested this on the Reykjanes dataset and it seems to work as expected.