fix(gaussian picker): Fix peak index finding; don't pick if only 1 sample above threshold
This commit fixes a bug which has been present in QM for some time - highlighted (not
caused) by the recent refactor. Previously, when finding the period of data that
exceeded the pick threshold, the final index was 1 sample too early, such that the last
sample of data above the threshold was lost when slicing by index. This has been fixed.

On top of this, the operation to add an additional pad of one sample at either end of this
period has been moved, and the variables re-named, to improve clarity.

And finally: previously, the gaussian fit was attempted even when only a single sample of
the onset function exceeded the threshold. This gives the minimum number of data points
to perform the optimisation; 3 (with the padding) for 3 parameters. The threshold has
now been increased to 2 samples above the threshold, and with the refactoring of the code
this choice is now explicit (should it need to be revisited in future).

Both of these changes will cause a slight difference to picking results. This has been
tested, and the changes compared to the previous benchmarks have been isolated to the
chosen revisions, which now lead to what has always been the desired behaviour.

There are also a couple of internal changes: the ValueError thrown by `curve_fit` when
it is fed too few data samples to perform the optimisation is now caught and passed to
a specific debug message. But hopefully with these most recent changes we should
no longer encounter this anyway.