fix(data.py): Temporarily switch to filtered data for times() function
The utility times() function serves up the times for each data sample, using the
functionality built into `obspy.Trace`. This had been changed to operate on
`data.waveforms`, which is not safe (highlighted in the review of PR #103). As
a temporary stop-gap, this has been changed to operating on
`data.filtered_waveforms`, to which data is only added after being resampled (if
necessary) and checked against the chosen data availability requirements. As such
it should function reliably as intended. A more permanent fix will be implemented
shortly (if required).