fix(gaussian picker): Revert to midtime between P & S where windows overlap
Where the pick windows for P and S phases (or generic phases <a> and <b>..) overlap,
they need to be separated at some intermediate point so that the picks are made in
the correct order. In the refactor, this was changed to the midpoint between the
calculated P pick window end and S pick window start, whereas previously it was
taken as the point mid-way between the modelled phase arrival times.

This is a fairly arbitrary decision, but on balance we have decided to stick with the
previous behaviour. This is because where the pick windows overlap, it is generally
because the traveltime is very small (i.e. the station is close to the hypocentre). The
biggest source of uncertainty in the predicted phase arrival times is then the uncertainty
in the origin time and hypocentre location of the event, rather than uncertainty in
the velocity model. As such, we aim to get as close to a full marginal window's worth
of window for each phase = midpoint between the modelled phase arrivals.