refactor(amplitudes): Make amplitudes compatible with phase agnosticism changes
Two changes made in this PR affect how amplitudes works.

1) The ability to make LUT's which do not contain traveltimes for both P and S

2) No longer writing a line to the picks file / dataframe when data is not available
   for that station/phase.

The former change has been handled by adding a new custom exception; to calculate
local magnitudes, the user's LUT must contain traveltimes for both P and S phases.
This rigidity may be reviewed in future.

The latter has been handled by changing how pick times are read in from the picks
dataframe.

On top of this, the channel codes iterated over have been improved to allow for
components labelled as "1" and "2" as well as "N" and "E", and a small bug has
been fixed which could lead to an UnboundLocalError exception.