fix(event summary): make phase agnostic
Make the event summary plotting phase agnostic - it was previously
hardcoded to assume there would be P and S phases available in the LUT.
This is obviously no longer the case and the code has been updated to
use more general expressions and use information on the actual phases
that are available.

Tested against:
- Icequake example, no changes
- Icequake example, dropping out some vertical and horizontal data from
a subset of stations
- Icequake example, only P phase. Make the LUT with just P and run
everything. A corollary of this is shows that using P alone (admittedly
without any tuning) does a significantly worse job!
- Icequake example, only S phase. Same as above - this is probably just
a paranoid edge case check, but it also works. Same lesson as above,
though - using P AND S improves everything significantly.

This is a quite minor fix. There will be a short follow up commit that
will address the hard-coded component codes (ZNE) - again, can make use
of the obspy `Stream.select()` method.