fix/feat(availability): handle old style files
Handle old-style availability file format. All of the required
information is there, as an availability of 1 in the old format is only
possible if all the required data was available for P AND S.

There is also an optional keyword that can be used to permanently convert
the availability files to the new style. For simplicity, this has to be run
externally to a QuakeMigrate run.

This means the benchmark test will now pass again (though this reqired a
slight adjustment to the test script - this was not included as the
benchmarks themselves will be updated in the near future).

Also adapted the availability file reading to use the better while loop
system (used in reading waveform data and triggered events). Though it
does raise the question - do we want the while test to be <= or <? I
have left it as <=, but I think it might lead to unhelpful "No
.StationAvailability file ..." messages in some edge cases.