feat(trigger plot): Multi-phase availablity plotting
Where multiple phases are used, adding the availability dataframe by row would
no longer give the 'number of stations available'. E.g. if 2 phases, and all
available, it would give stations * 2. We could either change labelling to #
of phases available, or split.

In this commit the feature is added to plot the phases separately. This only
supports 2 phases, currently; P and 1 other. Where the availablity is
identical for both phases it will revert to the previous plotting style. If more
than 2 phases are fed it will also do this.

A couple of other minor things: Add an np.ceil() to the ylim max, to avoid missing
the availability trace off the top of the figure where a small number of stations
are being used; remove an unused import.

Tested on the included examples, and on cases where only P or only S are used. All
seems to behave as expected.