feat/fix(availability): Add channel_count / n_channels kwargs for all_channels=True
If the user specifies all_channels=True, we need to check not just that all present
channels meet the availability criteria, but that all requested channels are present.
The most transparent way of doing this is by adding another onset class attribute
similar to `channel_maps`, but specifying the number of channels that should be
used to calculate the onset function for each phase. These are set by default to
`{"P" : 1, "S": 2}` and we don't foresee many people needing to interact with this.

This is passed to the `data.check_availability` method as n_channels, and this is
checked as an additional filter, if all present channels are 'available' and
`all_channels=True`. If `all_channels=True` `n_channels` is not specified, a
TypeError is raised.

There are also some minor style changes in response to the code review from @hemmelig.