feat(onset): Make phases an attribute of onset; more flexible phase choice
In the current implementation of 'phase agnosticism' the phases to be used for migration/
picking are decided by the phases present in the LUT. Here this is improved; the phases
to be used in any given migration and/or picking run can be selected by adjusting
`onset.phases`. The phases listed here will have to also be present in the LUT to be used
for these two purposes, but it will now be possible to experiment without calculating
a purpose-built LUT for each.

A couple more commits will follow to fully handle the eventualities this may throw up
in terms of plotting.