refactor(scan_rate): Hide scan_rate, but keep as internal variable
The `scan_rate` attribute has been added to QuakeScan in place of `sampling_rate`
in order to distinguish the input data / onset function sampling rate from the
sampling rate at which the 4D coalescence function is computed. However, we have
not yet finished refactoring the machinery of migrate() to make it possible for
these rates to be different.

For the time being, `scan_rate` will be retained as an internal attribute looking
forward to this new feature, but will not be exposed to the user (a later commit
will remove it from the template scripts). Furthermore, a message has been added
to advise the user of this constraint; their attempt to set it will not have any
effect, as the `scan_rate` will be read from the onset function `sampling_rate`
attribute internally.