fix(availability): Re-instate accommodation of off-sample timestamped data
Previously we checked that data covered the full timespan requested by comparing the
number of samples to the number required. This allowed data which was timestamped off-sample
(e.g. 50 Hz data sampled at 00:00:00.005, 00:00:00.025 etc.) to be included. This had
been changed to a stricter and explicit start and endtime check, but this commit rolls it
back to the previous "softer" behaviour.

Note, however, that this timing shift will NOT currently be taken into account when migrating
the data or making phase picks. As it is sub-sample it should be negligible in the scheme of
things, but the user should be aware of this.