fix/feat(stalta): `allow_gaps` clearly marked as experimental; fix fill_value
The `allow_gaps` and `full_timespan` kwargs for onset() and the check_availability method
were introduced to start to relax our rigid rejection of anything other than perfect data,
to allow (amongst other things) for the possible inclusion of event data for migration.

However, this is still an experimental feature: now clearly marked as such. A couple of
fixes are made in this commit which should leave it stable, though not guaranteed to work
as desired without some care and quality checking of input data on the part of the user.

Waveforms with gaps or which do not fill the timespan are now tapered and merged before
padding the start and end values. The taper is required to remove transients from filtering
which are more prone to happen where the trace is not so long. This will only affect the
padded periods for continuous data. The merge is required to join traces for gappy data.
In both the merge and pad the fill_value has been changed from 0 -> tiny float to avoid
floating point errors in onset calculation.