fix(amplitudes): Handle pick order issues from single-phase picking
Where an onset was only available for one phase, gaussianpicker does not check for
pick window overlap. This perhaps should be added in future, but for now if no pick
line is present in the .picks file we assume the pick hasn't been checked for whether
it is the corect side of the modelled arrival time for the other phase, and set both
to the modelled arrival times.

In this commit we also re-instate the previous behaviour of "is_picked" - this value is
True if _either_ phase was picked. It had been mistakenly changed to both.