feat(OnsetData): Add object on which to store onset calculation outputs
This brings the workflow for onset functions in line with that for how seismic data
is read from the waveform archive, then stored on a `WaveformData` object. It will
be used to store the calculated onset functions, a dictionary describing whether
onset functions are available for each station and phase, and the filtered and/or
resampled or otherwise pre-processed waveforms generated during onset function
calculation - primarily for use in plotting.