feat(onsets): Use `OnsetData` to store onsets and filtered waveforms
This follows the introduction of the `OnsetData` object in the previous commit. The
changes are mainly in where data is stored in stalta.py, how this is passed between
functions in scan.py and GaussianPicker, and how this is stored on the Event object.

There has also been a thorough clean of docstrings in each module, partly related to
this change, and partly long-overdue tidying / updating according to previous changes.