feat/fix(event plot): Use `OnsetData`, safer times(), docs
Main change is to use the `OnsetData` object to access filtered waveforms - this
follows changes in the previous commit (the call to this method in scan.py was
already updated there).

Also, generate times() for each trace individually, so that off-sample data is
plotted at the "correct" time. And don't extend the waveform gather plot past
the end of the waveforms (can happen if the event is at an extreme corner of the
grid).

And a big clean of docstrings.