feat(summary plot): Fully support using a subset of stations from the LUT
The package can now support using a subset of stations (as specified by the station
file) from the LUT - wahoo! This minor tweak removes one minor niggle with plotting,
in that case, by allowing the traveltimes for a subset of stations from the LUT to
be requested easily. (Previously an empty line with modelled arrivals was included
on the waveform gather, because the LUT served up traveltimes for all stations).