chore(update jnotebooks): Update jupyter notebook examples
This mirrors the changes made to the example scripts in the previous commit, plus some
more changes that had slipped through the net for a while.

Plus, some changes to language etc. to hopefully make the examples easier to run and
clearer to understand.

The change to plot/trigger.py is a stop-gap until #104 is merged. However: we need to check
that the way that change is implemented prevents the figures always showing up in the
notebooks.