fix(amplitude): Fix `filter_gain` bug
If both P and S windows didn't pass the data quality checks, `filter_gain` would
not be assigned a value, so if the noise measurement was made correctly, this
variable would attempt to be used when it was not set. This has been fixed.

The flatline check has also been added to the noise window measurement.