feat/fix(LUT plot): Only plot stations used for detect/locate
A previous commit dealt with the case where a subset of stations from the LUT were being
used in locate(), and removed the stations not being used from the waveform gather. This
commit does the same job but for the LUT plots (maps and cross sections) in the trigger
summary plot and event summary plot.

A kwarg is added to `lut.plot()` so that a list of stations can be provided; this subset
of stations is then selected from the LUT and plotted. These kwargs are passed into the
event summary plot (from event.data.stations == the stations specified by the user), and
read from the station availability data for the trigger summary plot.

This introduces the possibility to only plot stations which were available during the time
period of trigger; potentially desirable to get an accurate reflection of the geometry
of the active network. This has been added as a kwarg, defaulting to plotting all stations
in the station availability data, as before.

Plus some minor docstring cleanup, and adding the `station_data` attribute to `LUT` at
instantiation.